♻️ refactor(memory): move providers to internal/memory/{simple,lcm}/
Move simple and lcm provider implementations from plugins/memory/ to
internal/memory/. Drop plugin.go registration files — providers will be
constructed via a factory function. Add TODO for message conversion
deduplication.
