+++ internal/agent/pool_test.go
@@ -10,9 +10,9 @@ import (
 
 	appdb "github.com/CherryHQ/stella/internal/db"
 	"github.com/CherryHQ/stella/internal/memory"
+	lcmmemory "github.com/CherryHQ/stella/internal/memory/lcm"
 	"github.com/CherryHQ/stella/pkg/ai"
 	pkgplugins "github.com/CherryHQ/stella/pkg/plugins"
-	lcmmemory "github.com/CherryHQ/stella/plugins/memory/lcm"
 )
 
 // testMemoryProvider creates an LCM memory provider backed by an in-memory SQLite DB.
