+++ internal/memory/lcm/assembler.go
