+++ internal/memory/lcm/compaction.go
