+++ internal/memory/lcm/lcm_helpers_test.go
