+++ internal/memory/lcm/provider_extra_test.go
@@ -10,8 +10,8 @@ import (
 
 	appdb "github.com/CherryHQ/stella/internal/db"
 	"github.com/CherryHQ/stella/internal/memory"
+	"github.com/CherryHQ/stella/internal/memory/lcm"
 	"github.com/CherryHQ/stella/pkg/ai"
-	"github.com/CherryHQ/stella/plugins/memory/lcm"
 )
 
 func newLCMTestDB(t *testing.T) *sql.DB {
