+++ internal/memory/lcm/review.go
