+++ internal/memory/lcm/sessions.go
