+++ internal/memory/simple/provider_extra_test.go
@@ -8,8 +8,8 @@ import (
 
 	appdb "github.com/CherryHQ/stella/internal/db"
 	"github.com/CherryHQ/stella/internal/memory"
+	"github.com/CherryHQ/stella/internal/memory/simple"
 	"github.com/CherryHQ/stella/pkg/ai"
-	"github.com/CherryHQ/stella/plugins/memory/simple"
 )
 
 func newTestDB(t *testing.T) (*simple.Provider, func()) {
