--- plugins/memory/simple/plugin.go
@@ -1,35 +1,0 @@
-package simple
-
-import (
-	"context"
-	"errors"
-
-	"github.com/CherryHQ/stella/internal/memory"
-	pkgplugins "github.com/CherryHQ/stella/pkg/plugins"
-)
-
-func init() {
-	pkgplugins.Register("memory/simple", pkgplugins.PluginFunc(func(host pkgplugins.Host) {
-		host.SetInfo(pkgplugins.PluginInfo{
-			ID:           "memory/simple",
-			Kind:         "memory",
-			Name:         "simple",
-			DisplayName:  "Simple Sliding Window",
-			Description:  "Sliding-window memory without compaction or summaries.",
-			AdminVisible: true,
-			Capabilities: []string{
-				pkgplugins.CapabilityMemory,
-			},
-		})
-		host.AddMemory(pkgplugins.MemorySpec{
-			PluginID: "memory/simple",
-			Name:     "simple",
-			Build: func(ctx context.Context, build pkgplugins.MemoryContext) (memory.Provider, error) {
-				if build.DB == nil {
-					return nil, errors.New("simple plugin requires a shared DB connection")
-				}
-				return New(build.DB), nil
-			},
-		})
-	}))
-}
